Fix runtime error
